# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/public/screens/{instanceId}/heartbeat:
    post:
      tags:
      - screen-instance-controller
      operationId: heartbeat
      parameters:
      - name: instanceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenHeartbeatRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ScreenHeartbeatResponse"
...
  /api/admin/contests/{contestId}/exports/submissions.csv:
    get:
      tags:
      - data-export-controller
      operationId: submissions
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
  /api/admin/contests/{contestId}/exports/scoreboard.csv:
    get:
      tags:
      - data-export-controller
      operationId: scoreboard
      parameters:
      - name: contestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    ScreenPlaylistItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        targetView:
          type: string
          enum:
          - SCOREBOARD
          - FIRST_BLOOD
          - BALLOONS
          - FREEZE_COUNTDOWN
          - STATISTICS
          - RESOLVER
          - AWARDS
        durationSeconds:
          type: integer
          format: int32
        displayOrder:
          type: integer
          format: int32
...
    ScreenHeartbeatRequest:
      required:
      - clientToken
      - currentView
      type: object
      properties:
        clientToken:
          type: string
        currentView:
          type: string
          enum:
          - SCOREBOARD
          - FIRST_BLOOD
          - BALLOONS
          - FREEZE_COUNTDOWN
          - STATISTICS
          - RESOLVER
          - AWARDS
    ScreenGroupPlaybackResponse:
      type: object
      properties:
        groupId:
          type: integer
          format: int64
        groupName:
          type: string
        playlistId:
          type: integer
          format: int64
        loopEnabled:
          type: boolean
        status:
          type: string
          enum:
          - STOPPED
          - PLAYING
          - PAUSED
        startedAt:
          type: string
          format: date-time
        pausedElapsedSeconds:
          type: integer
          format: int64
        lockedView:
          type: string
          enum:
          - SCOREBOARD
          - FIRST_BLOOD
          - BALLOONS
          - FREEZE_COUNTDOWN
          - STATISTICS
          - RESOLVER
          - AWARDS
        version:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScreenPlaylistItemResponse"
    ScreenHeartbeatResponse:
      type: object
      properties:
        instanceId:
          type: integer
          format: int64
        serverTime:
          type: string
          format: date-time
        commandId:
          type: integer
          format: int64
        targetView:
          type: string
          enum:
          - SCOREBOARD
          - FIRST_BLOOD
          - BALLOONS
          - FREEZE_COUNTDOWN
          - STATISTICS
          - RESOLVER
          - AWARDS
        groupPlayback:
          $ref: "#/components/schemas/ScreenGroupPlaybackResponse"
...